requirements.txt
cryptography==3.4.8
pandas==1.3.4
python-dotenv==1.0.0
selenium==4.8.3
